feat(mcp): vanta mcp migration
